apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "structlog>=25.5.0",
    "prometheus-client>=0.23.1",
    "alembic>=1.18.4",
    "arq>=0.28.0",
    "asyncpg>=0.31.0",
    "email-validator>=2.2.0",
    "fastapi>=0.136.0",
    "google-api-python-client>=2.194.0",
    "google-auth>=2.49.2",
    "google-auth-oauthlib>=1.3.1",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.26",
    "redis>=5.3.1",
    "resend>=2.29.0",
    "sentry-sdk[fastapi]>=2.58.0",
    "sqlalchemy>=2.0.49",
    "sse-starlette>=3.3.4",
    "stripe>=15.0.1",
    "uvicorn[standard]>=0.44.0",
    "litellm>=1.83.10",
    "pyjwt[crypto]>=2.12.1",
    "svix>=1.90.0",
    "boto3>=1.42.91",
    "playwright>=1.49.0",
    "jinja2>=3.1.6",
    "cryptography>=46.0.7",
    "dnspython>=2.7.0",
    "icalendar>=6.1.1",
    "recurring-ical-events>=3.0.0",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "httpx-sse>=0.4.3",
    "mypy>=1.20.1",
    "psycopg[binary]>=3.3.3",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.11",
    "types-redis>=4.6.0.20241004",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.12"
files = ["app"]
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 120
target-version = "py312"
# Hand-written migrations are verbose; lint `alembic/env.py` only.
extend-exclude = ["alembic/versions", "app/seed_templates_data.py"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]

[tool.ruff.lint.per-file-ignores]
"scripts/*" = ["E402"]

